requirements.txt
asgiref==3.11.0
beautifulsoup4==4.14.3
Django==4.2.27
django-bootstrap-v5==1.0.11
pillow==12.1.0
python-decouple==3.8
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
